update
put it into the 'SEARCH_FUNCS'
