+++ camel/functions/__init__.py
@@ -15,11 +15,9 @@
 from .openai_function import OpenAIFunction
 from .math_functions import MATH_FUNCS
 from .search_functions import SEARCH_FUNCS
-from .web_search import WEB_FUNCS
 
 __all__ = [
     'OpenAIFunction',
     'MATH_FUNCS',
     'SEARCH_FUNCS',
-    'WEB_FUNCS',
 ]
